+++ core/src/builtin_widgets/svg.rs
@@ -27,7 +27,7 @@ pub mod named_svgs {
   const DEFAULT_SVG_KEY: &str = "__RIRBIR_DEFAULT_SVG__";
 
   static SVGS: LazyLock<Mutex<ahash::AHashMap<&'static str, Svg>>> = LazyLock::new(|| {
-    let svg = include_crate_svg!("src/builtin_widgets/default_named.svg", true, false);
+    let svg = asset!("./default_named.svg", "svg", inherit_fill = true, inherit_stroke = false);
     let mut set = ahash::AHashMap::new();
     set.insert(DEFAULT_SVG_KEY, svg);
     Mutex::new(set)
@@ -56,7 +56,7 @@ pub mod named_svgs {
   pub fn reset() {
     SVGS.lock().unwrap().clear();
 
-    let svg = include_crate_svg!("src/builtin_widgets/default_named.svg", true, false);
+    let svg = asset!("./default_named.svg", "svg", inherit_fill = true, inherit_stroke = false);
     named_svgs::register(DEFAULT_SVG_KEY, svg);
   }
 }